pyproject.toml
[project]
name = "slackbot"
dynamic = ["version"]
description = "An LLM bot for Slack"
authors = [
    { name = "Joyce Chung", email = "joycechungyt@gmail.com" },
    { name = "Nicholas Dykema", email = "nicowork13@gmail.com" },
    { name = "Elliott Steer", email = "essteer@pm.me" },
]
dependencies = [
    "langchain==0.1.13",
    "langchain-together==0.0.2.post1",
    "python-dotenv==1.0.1",
    "slack==0.0.2",
    "slack-bolt==1.18.1",
    "slackclient==2.9.4",
    "together==0.2.11",
    "torch==2.2.2",
    "transformers==4.39.2",
]
requires-python = ">=3.9"
license = { file = "LICENSE" }
readme = "README.md"
keywords = ["LLM", "Slack", "Bots"]
classifiers = [
    "Operating System :: OS Independent",
    "Natural Language :: English",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License 2.0",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
documentation = "https://github.com/JENX-AI/slackbot/blob/main/README.md"
repository = "https://github.com/JENX-AI/slackbot"
issues = "https://github.com/JENX-AI/slackbot/issues"
